Moved a component to molecules from atoms
